requirements.txt
aiogram==3.13.1
gspread==6.1.2
oauth2client==4.1.3
pandas==2.2.3
pydantic==2.9.2
pydantic_settings==2.5.2
python-decouple==3.8
python-dotenv==1.0.1
